+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v4.3.0
+
+- Alternative scoring: formula is solved / (total - added)
+- Fix homepage when no screenshot added yet
+- Fix approve / disapprove already moderated screenshots
+- Fix account page infinite loader when no screenshot solved / added
+
 ## v4.2.0
 
 - Better random screenshots: you see all unseen screenshots without repetition (with 200 unseen sc max)